migrate meme-card
